refactor(Estimate): Moved common code into trait
Moved common code for invoice and estimate into a common trait
