+++ src/Request/Data/Invoice/InvoiceCreateData.php
@@ -9,10 +9,12 @@
 namespace ZEROSPAM\Freshbooks\Request\Data\Invoice;
 
 use ZEROSPAM\Freshbooks\Request\Data\ArrayableData;
+use ZEROSPAM\Freshbooks\Request\Data\Common\BaseWritableFieldsTrait;
 
 class InvoiceCreateData extends ArrayableData
 {
-    use WritableInvoiceFieldsTrait;
+    use BaseWritableFieldsTrait,
+        WritableInvoiceFieldsTrait;
 
     /** @var int */
     private $ownerid;
